+++ src/compiler/checker/lines_0_2000.rs
@@ -430,74 +430,6 @@ impl TypeChecker {
 
         result
     }
-
-    pub(crate) fn resolve_alias(&self, symbol: Rc<Symbol>) -> Rc<Symbol> {
-        unimplemented!()
-    }
-
-    pub(crate) fn resolve_entity_name(
-        &self,
-        name: &Node, /*EntityNameOrEntityNameExpression*/
-        meaning: SymbolFlags,
-        ignore_errors: Option<bool>,
-        dont_resolve_alias: Option<bool>,
-    ) -> Option<Rc<Symbol>> {
-        let ignore_errors = ignore_errors.unwrap_or(false);
-        let dont_resolve_alias = dont_resolve_alias.unwrap_or(false);
-        if node_is_missing(name) {
-            return None;
-        }
-
-        let symbol: Option<Rc<Symbol>>;
-        match name {
-            Node::Expression(Expression::Identifier(name)) => {
-                let message = if false {
-                    unimplemented!()
-                } else {
-                    self.get_cannot_find_name_diagnostic_for_name(&*get_first_identifier(name))
-                };
-                let symbol_from_js_prototype: Option<Rc<Symbol>> =
-                    if false { unimplemented!() } else { None };
-                symbol = self.get_merged_symbol(self.resolve_name(
-                    Some(name),
-                    &name.escaped_text,
-                    meaning,
-                    if ignore_errors || symbol_from_js_prototype.is_some() {
-                        None
-                    } else {
-                        Some(message)
-                    },
-                    Some(name.node_wrapper()),
-                    true,
-                    Some(false),
-                ));
-                if symbol.is_none() {
-                    unimplemented!()
-                }
-            }
-            // else if name.kind() == SyntaxKind::QualifiedName
-            //     || name.kind() == SyntaxKind::PropertyAccessExpression
-            //     unimplemented!()
-            _ => Debug_.assert_never(name, Some("Unknown entity name kind.")),
-        }
-        let symbol = symbol.unwrap();
-        if symbol.flags().intersects(meaning) || dont_resolve_alias {
-            Some(symbol)
-        } else {
-            Some(self.resolve_alias(symbol))
-        }
-    }
-
-    pub(crate) fn get_merged_symbol(&self, symbol: Option<Rc<Symbol>>) -> Option<Rc<Symbol>> {
-        symbol
-    }
-
-    pub(crate) fn get_symbol_of_node<TNode: NodeInterface>(
-        &self,
-        node: &TNode,
-    ) -> Option<Rc<Symbol>> {
-        self.get_merged_symbol(node.maybe_symbol())
-    }
 }
 
 fn create_node_builder() -> NodeBuilder {
