+++ src/compiler/checker/lines_2000_4000.rs
@@ -1,0 +1,106 @@
+#![allow(non_upper_case_globals)]
+
+use bitflags::bitflags;
+use std::borrow::Borrow;
+use std::cell::{Ref, RefCell, RefMut};
+use std::collections::{HashMap, HashSet};
+use std::ptr;
+use std::rc::Rc;
+
+use crate::{
+    TypeNode, UnionOrIntersectionType, UnionOrIntersectionTypeInterface, UnionType,
+    VariableDeclaration, VariableStatement, __String, bind_source_file, chain_diagnostic_messages,
+    create_diagnostic_collection, create_diagnostic_for_node,
+    create_diagnostic_for_node_from_message_chain, create_printer, create_symbol_table,
+    create_text_writer, declaration_name_to_string, escape_leading_underscores, every, factory,
+    first_defined, first_or_undefined, for_each, get_effective_initializer,
+    get_effective_type_annotation_node, get_first_identifier, get_name_of_declaration,
+    get_object_flags, get_source_file_of_node, get_synthetic_factory, has_dynamic_name,
+    has_initializer, is_binding_element, is_external_or_common_js_module, is_identifier_text,
+    is_object_literal_expression, is_private_identifier, is_property_assignment,
+    is_property_declaration, is_property_signature, is_variable_declaration, node_is_missing,
+    object_allocator, unescape_leading_underscores, using_single_line_string_writer, ArrayTypeNode,
+    BaseInterfaceType, BaseIntrinsicType, BaseLiteralType, BaseNodeFactorySynthetic,
+    BaseObjectType, BaseType, BaseUnionOrIntersectionType, CharacterCodes, Debug_, Diagnostic,
+    DiagnosticCollection, DiagnosticMessage, DiagnosticMessageChain, Diagnostics, EmitHint,
+    EmitTextWriter, Expression, ExpressionStatement, FreshableIntrinsicType,
+    HasExpressionInitializerInterface, InterfaceDeclaration, InterfaceType, IntrinsicType,
+    KeywordTypeNode, LiteralLikeNode, LiteralLikeNodeInterface, LiteralType, LiteralTypeInterface,
+    NamedDeclarationInterface, Node, NodeInterface, Number, NumberLiteralType, NumericLiteral,
+    ObjectFlags, ObjectFlagsTypeInterface, ObjectLiteralExpression, PrefixUnaryExpression,
+    PrinterOptions, PropertyAssignment, PropertySignature, RelationComparisonResult,
+    ResolvableTypeInterface, ResolvedTypeInterface, SourceFile, Statement, StringLiteralType,
+    Symbol, SymbolFlags, SymbolFormatFlags, SymbolTable, SymbolTracker, SyntaxKind, Ternary, Type,
+    TypeChecker, TypeCheckerHost, TypeElement, TypeFlags, TypeInterface, TypeReferenceNode,
+    UnionReduction, VariableLikeDeclarationInterface,
+};
+
+impl TypeChecker {
+    pub(crate) fn resolve_alias(&self, symbol: Rc<Symbol>) -> Rc<Symbol> {
+        unimplemented!()
+    }
+
+    pub(crate) fn resolve_entity_name(
+        &self,
+        name: &Node, /*EntityNameOrEntityNameExpression*/
+        meaning: SymbolFlags,
+        ignore_errors: Option<bool>,
+        dont_resolve_alias: Option<bool>,
+    ) -> Option<Rc<Symbol>> {
+        let ignore_errors = ignore_errors.unwrap_or(false);
+        let dont_resolve_alias = dont_resolve_alias.unwrap_or(false);
+        if node_is_missing(name) {
+            return None;
+        }
+
+        let symbol: Option<Rc<Symbol>>;
+        match name {
+            Node::Expression(Expression::Identifier(name)) => {
+                let message = if false {
+                    unimplemented!()
+                } else {
+                    self.get_cannot_find_name_diagnostic_for_name(&*get_first_identifier(name))
+                };
+                let symbol_from_js_prototype: Option<Rc<Symbol>> =
+                    if false { unimplemented!() } else { None };
+                symbol = self.get_merged_symbol(self.resolve_name(
+                    Some(name),
+                    &name.escaped_text,
+                    meaning,
+                    if ignore_errors || symbol_from_js_prototype.is_some() {
+                        None
+                    } else {
+                        Some(message)
+                    },
+                    Some(name.node_wrapper()),
+                    true,
+                    Some(false),
+                ));
+                if symbol.is_none() {
+                    unimplemented!()
+                }
+            }
+            // else if name.kind() == SyntaxKind::QualifiedName
+            //     || name.kind() == SyntaxKind::PropertyAccessExpression
+            //     unimplemented!()
+            _ => Debug_.assert_never(name, Some("Unknown entity name kind.")),
+        }
+        let symbol = symbol.unwrap();
+        if symbol.flags().intersects(meaning) || dont_resolve_alias {
+            Some(symbol)
+        } else {
+            Some(self.resolve_alias(symbol))
+        }
+    }
+
+    pub(crate) fn get_merged_symbol(&self, symbol: Option<Rc<Symbol>>) -> Option<Rc<Symbol>> {
+        symbol
+    }
+
+    pub(crate) fn get_symbol_of_node<TNode: NodeInterface>(
+        &self,
+        node: &TNode,
+    ) -> Option<Rc<Symbol>> {
+        self.get_merged_symbol(node.maybe_symbol())
+    }
+}
